Move useServerStatus hook to AppLoading
